+++ pvlib/__init__.py
@@ -11,4 +11,4 @@ from pvlib import tracking
 from pvlib import pvsystem
 from pvlib import spa
 from pvlib import modelchain
-from pvlib import singlediode_methods
+from pvlib import singlediode
